better serialization

+++ src/emucore/M6532.cxx
@@ -334,7 +334,7 @@ bool M6532::save(Serializer& out)
     // Output the RAM
     out.putInt(128);
     for(uInt32 t = 0; t < 128; ++t)
-      out.putInt(myRAM[t]);
+      out.putByte(myRAM[t]);
 
     out.putInt(myTimer);
     out.putInt(myIntervalShift);
@@ -371,7 +371,7 @@ bool M6532::load(Deserializer& in)
     // Input the RAM
     uInt32 limit = (uInt32) in.getInt();
     for(uInt32 t = 0; t < limit; ++t)
-      myRAM[t] = (uInt8) in.getInt();
+      myRAM[t] = (uInt8) in.getByte();
 
     myTimer = (uInt32) in.getInt();
     myIntervalShift = (uInt32) in.getInt();
